Fix typo in package names
