+++ _internal/fg7/minecraft/1.13.2/src/main/java/net/minecraftforge/test/TestMod.java
@@ -1,0 +1,30 @@
+package net.minecraftforge.test;
+
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.gameevent.TickEvent;
+import net.minecraftforge.fml.server.ServerLifecycleHooks;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+
+@Mod("test")
+public class TestMod {
+    public TestMod() {
+        MinecraftForge.EVENT_BUS.addListener(this::onServerTick);
+    }
+
+    private void onServerTick(TickEvent.ServerTickEvent event) {
+        if (event.phase != TickEvent.Phase.END) return;
+        System.out.println("SERVER_TICK_SUCCESS");
+        try {
+            Files.write(new File("test_marker.txt").toPath(), "SERVER_TICK_SUCCESS".getBytes(StandardCharsets.UTF_8));
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        } finally {
+            ServerLifecycleHooks.getCurrentServer().initiateShutdown();
+        }
+    }
+}
